feat(stringCalculator) : refactoring(커스텀 예외 처리(NegativeNumberException) 추가)
