+++ src/main/java/calculator/StringCalculator.java
@@ -1,5 +1,7 @@
 package calculator;
 
+import exception.NegativeNumberException;
+
 import java.util.List;
 
 public class StringCalculator {
@@ -18,8 +20,11 @@ public class StringCalculator {
 
         // 2. 숫자 - 나눈 문자열을 숫자로 변환하기
         Numbers numbers = new Numbers(splits);
-        if (numbers.isNullOrEmpty() || numbers.hasNegativeNumber()) {
+        if (numbers.isNullOrEmpty()) {
             throw new RuntimeException("invalid numbers : " + numbers.getNumbers());
+        }
+        if (numbers.hasNegativeNumber()) {
+            throw new NegativeNumberException("negative numbers found : " + numbers.getNumbers());
         }
 
         // 3. 연산 - 숫자들의 합 구하기
