+++ src/main/java/exception/GlobalExceptionHandler.java
@@ -1,0 +1,22 @@
+package exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(NegativeNumberException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public String handleNegativeNumberException(NegativeNumberException ex) {
+        return ex.getMessage();
+    }
+
+    @ExceptionHandler(Exception.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public String handleGlobalException(Exception ex) {
+        return "Internal Server Error : " + ex.getMessage();
+    }
+}
