+++ src/main/java/exception/NegativeNumberException.java
@@ -1,0 +1,9 @@
+package exception;
+
+public class NegativeNumberException extends RuntimeException {
+
+    public NegativeNumberException(String message) {
+        super(message);
+    }
+
+}
